Refined a bit the communication protocol in order to not send useless informations to the client and not force it to write more data than it should know, fixed bugs in serialization of the structures

+++ src/server.h
@@ -22,16 +22,18 @@ struct socks {
     int serversock;
 };
 
+
 struct reply {
-    uint16_t type;
+    uint8_t type;
     uint8_t qos;
     int fd;
     char *data;
     char *channel;
 };
 
+
 struct global {
-    uint32_t next_id;
+    uint64_t next_id;
     map *channels;
     map *ack_waiting;
 };
